fix: fix bugs
