+++ api/favorite/service/v1/favorite.proto
@@ -24,11 +24,11 @@ service FavoriteService {
 
 message IsFavoriteRequest {
   uint32 user_id = 1; // 用户id
-  repeated uint32 video_ids = 3; // 视频id
+  repeated uint32 video_ids = 2; // 视频id
 }
 
 message IsFavoriteReply {
-  repeated bool is_favorite = 3; // true-已点赞，false-未点赞
+  repeated bool is_favorite = 1; // true-已点赞，false-未点赞
 }
 
 message FavoriteListRequest {
