+++ api/user/service/v1/user.proto
@@ -110,7 +110,7 @@ message UserInfosRequest {
 }
 
 message UserInfosReply {
-	repeated User users = 3; // 用户信息列表
+	repeated User users = 1; // 用户信息列表
 }
 
 message UpdateFollowRequest {
