+++ app/user/service/internal/conf/conf.proto
@@ -9,6 +9,7 @@ import "google/protobuf/duration.proto";
 message Bootstrap {
   Server server = 1;
   Data data = 2;
+  JWT jwt = 3;
 }
 
 message Server {
@@ -39,4 +40,15 @@ message Data {
   }
   Database database = 1;
   Redis redis = 2;
+}
+
+message JWT {
+  message Http {
+    string token_key = 1;
+  }
+  message Grpc {
+    string token_key = 1;
+  }
+  Http http = 1;
+  Grpc grpc = 2;
 }
